+++ cli-console/build.gradle.kts
@@ -16,6 +16,10 @@ dependencies {
   val jlineVersion = deps.getProperty("version.jline")
   implementation("org.jline", "jline-terminal-jansi", version = jlineVersion)
   implementation("org.jline", "jline-builtins", version = jlineVersion)
+  testImplementation("org.junit.jupiter", "junit-jupiter", version = deps.getProperty("version.junit"))
+  testImplementation("org.hamcrest", "hamcrest", version = deps.getProperty("version.hamcrest"))
+  testImplementation("org.ice1000.jimgui", "core", version = deps.getProperty("version.jimgui"))
+  // testImplementation("org.ice1000.jimgui", "fun", version = deps.getProperty("version.jimgui"))
 }
 
 plugins {
