+++ cli-console/src/test/java/org/aya/cli/console/ImGuiTrace.java
@@ -1,6 +1,6 @@
 // Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli;
+package org.aya.cli.console;
 
 import kala.collection.Seq;
 import kala.collection.immutable.ImmutableSeq;
