+++ cli-console/src/test/java/org/aya/cli/console/WidthPreviewToy.java
@@ -1,6 +1,6 @@
-// Copyright (c) 2020-2022 Tesla (Yinsen) Zhang.
+// Copyright (c) 2020-2023 Tesla (Yinsen) Zhang.
 // Use of this source code is governed by the MIT license that can be found in the LICENSE.md file.
-package org.aya.cli;
+package org.aya.cli.console;
 
 import org.aya.prettier.AyaPrettierOptions;
 import org.aya.pretty.doc.Doc;
